+++ APIBlox.AspNetCore.CommandsQueriesControllersOhMy/Controllers/DynamicPostController.cs
@@ -1,15 +1,18 @@
 ﻿using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using APIBlox.AspNetCore.ActionResults;
 using APIBlox.AspNetCore.Contracts;
 using APIBlox.AspNetCore.Types;
+using APIBlox.NetCore.Extensions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 
 namespace APIBlox.AspNetCore.Controllers
 {
-    /// <summary>
+ /// <summary>
     ///     Class DynamicPostController.
     /// </summary>
     /// <typeparam name="TRequest">The type of the t request.</typeparam>
@@ -25,14 +28,17 @@ namespace APIBlox.AspNetCore.Controllers
         where TResponse : IResource<TId>
     {
         private readonly ICommandHandler<TRequest, HandlerResponse> _createCommand;
+        private readonly ILogger<DynamicPostController<TRequest, TResponse, TId>> _log;
 
         /// <summary>
         ///     Initializes a new instance of the <see cref="DynamicPostController{TRequest, TResponse, TId}" /> class.
         /// </summary>
+        /// <param name="loggerFactory">LoggerFactory</param>
         /// <param name="createCommand">The create command.</param>
-        public DynamicPostController(ICommandHandler<TRequest, HandlerResponse> createCommand)
+        public DynamicPostController(ILoggerFactory loggerFactory, ICommandHandler<TRequest, HandlerResponse> createCommand)
         {
             _createCommand = createCommand;
+            _log = loggerFactory.CreateLogger<DynamicPostController<TRequest, TResponse, TId>>();
         }
 
         /// <summary>
@@ -63,7 +69,42 @@ namespace APIBlox.AspNetCore.Controllers
                     "When responding to a POST you must either set an error or pass some results!"
                 );
 
-            return errorResult ?? CreatedAtRoute(new {ret.Result.Id}, ret.Result);
+            if (!(errorResult is null))
+                return errorResult;
+
+            var id = FindId(ret.Result);
+
+            return id == -1
+                ? (IActionResult) Ok(ret.Result)
+                : (IActionResult) CreatedAtRoute(new {id}, ret.Result);
+        }
+
+        // But it seems sometimes responses do not match what was requested, I think that's a violation but what do I know...
+        private object FindId(object result)
+        {
+            try
+            {
+                var t = result.GetType();
+                var props = t.GetProperties();
+
+                var id = props.FirstOrDefault(p => p.Name.EqualsEx("Id"));
+
+                if (id is null)
+                {
+                    foreach (var pi in props)
+                        return FindId(t.GetProperty(pi.Name).GetValue(result, null));
+                }
+                else
+                    return t.GetProperty(id.Name).GetValue(result, null);
+            }
+            catch (Exception ex)
+            {
+                _log.LogError(() => $"Could not determine ID for CreatedAtRoute result!  Ex: {ex.Message}");
+
+                return -1;
+            }
+
+            return -1;
         }
     }
 }